pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "activity_predictor"
authors = [{name = "William Droz", email = "william.droz@idiap.ch"},]
readme = "README.md"
license = {file = "LICENSE"}
classifiers = ["License :: OSI Approved :: Apache Software License"]
dynamic = ["version", "description"]
dependencies = [
    "python-dotenv == 1.0.0",
    "requests == 2.28.2",
]
[project.urls]
Home = "https://gitlab.idiap.ch/wenet/activities-predictor"

[project.optional-dependencies]
dev = [
    "bandit >= 1.7.4,<2",
    "black == 22.3.0",
    "coverage >= 6.4.2,<7.0.0",
    "flake8 >=4.0.1,<5.0.0",
    "mypy >=0.971,<1.0",
    "pre-commit >=2.17.0,<3.0.0",
    "pytest >=7.1.2,<8.0.0",
    "ruff == 0.0.254",
]

[tool.ruff]
# Enable pycodestyle (`E`) and Pyflakes (`F`) codes by default.
select = ["E", "F"]
ignore = ["E501"]

# Allow autofix for all enabled rules (when `--fix`) is provided.
fixable = ["A", "B", "C", "D", "E", "F"]
unfixable = []

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Same as Black.
line-length = 88

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# Assume Python 3.10.
target-version = "py310"

[tool.ruff.mccabe]
# Unlike Flake8, default to a complexity level of 10.
max-complexity = 10
